RSPY-766: Refacto of rs-dpr-service part 2
