+++ rs_dpr_service/dask/dask_cluster_handler.py
@@ -1,0 +1,226 @@
+# Copyright 2025 CS Group
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Module to handle connection to Dask cluster."""
+
+import os
+
+from dask.distributed import (  # LocalCluster,
+    Client,
+)
+from dask_gateway import Gateway
+from dask_gateway.auth import BasicAuth, JupyterHubAuth
+
+from rs_dpr_service.dask import call_dask
+from rs_dpr_service.utils.logging import Logging
+from rs_dpr_service.utils.utils import env_bool
+
+logger = Logging.default(__name__)
+
+# True if the 'RSPY_LOCAL_MODE' environemnt variable is set to 1, true or yes (case insensitive).
+# By default: if not set or set to a different value, return False.
+LOCAL_MODE: bool = env_bool("RSPY_LOCAL_MODE", default=False)
+
+
+class DaskClusterHandler:
+    """Class to handle connection to Dask cluster"""
+
+    def __init__(self, cluster_address: str, cluster_name: str):
+        self.cluster_name = cluster_name
+        self.cluster_address = cluster_address
+        self.cluster = None
+
+    def dask_cluster_connect(self):  # pylint: disable=too-many-branches, too-many-statements, too-many-locals
+        """Connects a dask cluster scheduler
+        Establishes a connection to a Dask cluster, either in a local environment or via a Dask Gateway in
+        a Kubernetes cluster. This method checks if the cluster is already created (for local mode) or connects
+        to a Dask Gateway to find or create a cluster scheduler (for Kubernetes mode, see RSPY_LOCAL_MODE env var).
+
+        1. **Local Mode**:
+        - If `self.cluster` already exists, it assumes the Dask cluster was created when the application started,
+            and proceeds without creating a new cluster.
+
+        2. **Kubernetes Mode**:
+        - If `self.cluster` is not already defined, the method attempts to connect to a Dask Gateway
+            (using environment variables `DASK_GATEWAY__ADDRESS` and `DASK_GATEWAY__AUTH__TYPE`) to
+            retrieve a list of existing clusters.
+        - If no clusters are available, it attempts to create a new cluster scheduler.
+
+        Raises:
+            RuntimeError: Raised if the cluster name is None, required environment variables are missing,
+                        cluster creation fails or authentication errors occur.
+            KeyError: Raised if the necessary Dask Gateway environment variables (`DASK_GATEWAY__ADDRESS`,
+                `DASK_GATEWAY__AUTH__TYPE`, `RSPY_DASK_DPR_SERVICE_CLUSTER_NAME`, `JUPYTERHUB_API_TOKEN` ) are not set.
+            IndexError: Raised if no clusters are found in the Dask Gateway and new cluster creation is attempted.
+            dask_gateway.exceptions.GatewayServerError: Raised when there is a server-side error in Dask Gateway.
+            dask_gateway.exceptions.AuthenticationError: Raised if authentication to the Dask Gateway fails.
+            dask_gateway.exceptions.ClusterLimitExceeded: Raised if the limit on the number of clusters is exceeded.
+
+        Behavior:
+        1. **Cluster Creation and Connection**:
+            - In Kubernetes mode, the method tries to connect to an existing cluster or creates
+            a new one if none exists.
+            - Error handling includes catching issues like missing environment variables, authentication failures,
+            cluster creation timeouts, or exceeding cluster limits.
+
+        2. **Logging**:
+            - Logs the list of available clusters if connected via the Dask Gateway.
+            - Logs the success of the connection or any errors encountered during the process.
+            - Logs the Dask dashboard URL and the number of active workers.
+
+        3. **Client Initialization**:
+            - Once connected to the Dask cluster, the method creates a Dask `Client` object for managing tasks
+            and logs the number of running workers.
+            - If no workers are found, it scales the cluster to 1 worker.
+
+        4. **Error Handling**:
+            - Handles various exceptions during the connection and creation process, including:
+            - Missing environment variables.
+            - Failures during cluster creation.
+            - Issues related to cluster scaling, worker retrieval, or client creation.
+            - If an error occurs, the method logs the error and attempts to gracefully handle failure.
+
+        Returns:
+            Dask client
+        """
+
+        # If self.cluster is already initialized, it means the application is running in local mode, and
+        # the cluster was created when the application started.
+
+        # Connect to the gateway and get the list of the clusters
+        try:
+            # In local mode, authenticate to the dask cluster with username/password
+            if LOCAL_MODE:
+                gateway_auth = BasicAuth(
+                    os.environ["LOCAL_DASK_USERNAME"],
+                    os.environ["LOCAL_DASK_PASSWORD"],
+                )
+
+            # Cluster mode
+            else:
+                # check the auth type, only jupyterhub type supported for now
+                auth_type = os.environ["DASK_GATEWAY__AUTH__TYPE"]
+                # Handle JupyterHub authentication
+                if auth_type == "jupyterhub":
+                    gateway_auth = JupyterHubAuth(api_token=os.environ["JUPYTERHUB_API_TOKEN"])
+                else:
+                    logger.error(f"Unsupported authentication type: {auth_type}")
+                    raise RuntimeError(f"Unsupported authentication type: {auth_type}")
+
+            gateway = Gateway(
+                address=self.cluster_address,
+                auth=gateway_auth,
+            )
+
+            # Sort the clusters by newest first
+            clusters = sorted(gateway.list_clusters(), key=lambda cluster: cluster.start_time, reverse=True)
+            logger.debug(f"Cluster list for gateway {self.cluster_address!r}: {clusters}")
+
+            # In local mode, get the first cluster from the gateway.
+            cluster_id = None
+            if LOCAL_MODE:
+                if clusters:
+                    cluster_id = clusters[0].name
+
+            # In cluster mode, get the identifier of the cluster whose name is equal to the cluster_name variable.
+            # Protection for the case when this cluster does not exit
+            else:
+                logger.info(f"my cluster name: {self.cluster_name}")
+
+                for cluster in clusters:
+                    logger.info(f"Existing cluster names: {cluster.options.get('cluster_name')}")
+
+                    is_equal = cluster.options.get("cluster_name") == self.cluster_name
+                    logger.info(f"Is equal: {is_equal}")
+
+                cluster_id = next(
+                    (
+                        cluster.name
+                        for cluster in clusters
+                        if isinstance(cluster.options, dict)
+                        and cluster.options.get("cluster_name") == self.cluster_name
+                    ),
+                    None,
+                )
+                logger.info(f"Cluster id: {cluster_id}")
+
+            if not cluster_id:
+                raise IndexError(f"Dask cluster with 'cluster_name'={self.cluster_name!r} was not found.")
+
+            self.cluster = gateway.connect(cluster_id)
+            if not self.cluster:
+                logger.exception("Failed to create the cluster")
+                raise RuntimeError("Failed to create the cluster")
+            logger.info(f"Successfully connected to the {self.cluster_name} dask cluster")
+
+            # This cluster id is needed by the eopf dask scheduler to connect later to this cluster.
+            # This is something like "dask-gateway.17e196069443463495547eb97f532834"
+            os.environ["DASK_CLUSTER_EOPF_NAME"] = cluster_id
+
+        except KeyError as e:
+            logger.exception(
+                "Failed to retrieve the required connection details for "
+                "the Dask Gateway from one or more of the following environment variables: "
+                "DASK_GATEWAY__ADDRESS, RSPY_DASK_DPR_SERVICE_CLUSTER_NAME, "
+                f"JUPYTERHUB_API_TOKEN, DASK_GATEWAY__AUTH__TYPE. {e}",
+            )
+
+            raise RuntimeError(
+                f"Failed to retrieve the required connection details for Dask Gateway. Missing key:{e}",
+            ) from e
+        except IndexError as e:
+            logger.exception(f"Failed to find the specified dask cluster: {e}")
+            raise RuntimeError(f"No dask cluster named '{self.cluster_name}' was found.") from e
+
+        logger.debug("Cluster dashboard: %s", self.cluster.dashboard_link)
+        # create the client as well
+        client = Client(self.cluster)
+
+        # Forward logging from dask workers to the caller
+        client.forward_logging()
+
+        # Upload local module to the dask client.
+        call_dask.upload_this_module(client)
+
+        def set_dask_env(host_env: dict):
+            """Pass environment variables to the dask workers."""
+            for name in ["S3_ACCESSKEY", "S3_SECRETKEY", "S3_ENDPOINT", "S3_REGION"]:
+                os.environ[name] = host_env[name]
+
+            # Some kind of workaround for boto3 to avoid checksum being added inside
+            # the file contents uploaded to the s3 bucket e.g. x-amz-checksum-crc32:xxx
+            # See: https://github.com/boto/boto3/issues/4435
+            os.environ["AWS_REQUEST_CHECKSUM_CALCULATION"] = "when_required"
+            os.environ["AWS_RESPONSE_CHECKSUM_VALIDATION"] = "when_required"
+
+        client.run(set_dask_env, os.environ)
+
+        # This is a temporary fix for the dask cluster settings which does not create a scheduler by default
+        # This code should be removed as soon as this is fixed in the kubernetes cluster
+        try:
+            logger.debug(f"{client.get_versions(check=True)}")
+            workers = client.scheduler_info()["workers"]
+            logger.info(f"Number of running workers: {len(workers)}")
+
+        except Exception as e:  # pylint: disable=broad-exception-caught
+            logger.exception(f"Dask cluster client failed: {e}")
+            raise RuntimeError(f"Dask cluster client failed: {e}") from e
+        if len(workers) == 0:
+            logger.info("No workers are currently running in the Dask cluster. Scaling up to 1.")
+            self.cluster.scale(1)
+
+        # Check the cluster dashboard
+        logger.debug(f"Dask Client: {client} | Cluster dashboard: {self.cluster.dashboard_link}")
+
+        return client
